Add ForiegnKey on_delete for Django compat
The old migrations are given the default on_delete CASCADE,
otherwise the current migrations fail to load on modern Django
with an Exception due to the missing on_delete.

A new migration is added to switch the on_delete to prevent
deletion of logs when a host or script is deleted.
The deletion of a script is blocked if it was used in a log,
however the host can be deleted and its log will loose
the reference to which host it was run on.  In most cases
the host is not critical information.

A new squashed migration is added due to the amount of changes.